requirements.txt
ccxt
pandas
numpy
talib
requests
openai
google-cloud-logging
google-cloud-monitoring